Fix test_orc_fallback for Spark400 (#12908)
contributes to https://github.com/NVIDIA/spark-rapids/issues/11013

### bug description
This case `test_orc_fallback` is testing the fallback to CPU behavior
for configs:
- `spark.rapids.sql.format.orc.enabled'
- 'spark.rapids.sql.format.orc.read.enabled

On Spark400, fallback works, but the addition operator throws exception
because Spark400's default Ansi mode is true.

### fix
Change the `select a + b` to `select a, b` to avoid overflow which
causes exception for Ansi mode.

Signed-off-by: Chong Gao <res_life@163.com>

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>