Fix Spark 4.0 IT failures: csv_test.py (#12946)
closes https://github.com/NVIDIA/spark-rapids/issues/11016

### latest failed cases

- test_csv_fallback
- test_read_valid_and_invalid_dates

### fix
- test_csv_fallback
Spark 4.0 uses ANSI by default, and this case fallback to CPU, then CPU
catches some invalid value and causes error.
Using non-ANSI mode will fix this.

- test_read_valid_and_invalid_dates
Refer to issue:  https://github.com/NVIDIA/spark-rapids/issues/12943
The fix is the same with
https://github.com/NVIDIA/spark-rapids/pull/12945, change the error from
"DateTimeException" to "Exception". This is a temparary fix. The
[#12943](https://github.com/NVIDIA/spark-rapids/issues/12943) is
tracking the issue.

Signed-off-by: Chong Gao <res_life@163.com>

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>