Fix Spark 4.0 IT failures: test_conv (#12947)
closes https://github.com/NVIDIA/spark-rapids/issues/11030


### Failed cases
- test_conv_with_more_valid_values
- test_conv_with_more_invalid_values

Refer to
[link](https://github.com/NVIDIA/spark-rapids/issues/11030#issuecomment-2971850822)

### Fixes
- test_conv_with_more_valid_values
The regexp `r'-?[0-9a-zA-Z]{1,15}'` is wrong, 15 digits number of 36
radix exceeds max long value which is invalid:
36^15 > Long.MaxValue
Fix: Change the number 15 to 11, which generates values are less than
max long.

- test_conv_with_more_invalid_values
This is testing invalid values will get null for non-ANSI mode.
So for Spark 4.0, should explicitly specify non-ANSI mode because
default ANSI mode for Spark 4.0 is true.

Signed-off-by: Chong Gao <res_life@163.com>

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>