Xfail some tests in json_matrix_test.py for Spark 400+ (#12957)
close https://github.com/NVIDIA/spark-rapids/issues/12936

The failure is because of the whitespace differences in the output JSON
string. e.g.
```
-Row(data='{"numeric": 123, "text": "example", "flag": true, "details": { "timestamp": "2024-08-22T10:00:00Z","list": [1, 2, 3]}}')
+Row(data='{"numeric":123,"text":"example","flag":true,"details":{"timestamp":"2024-08-22T10:00:00Z","list":[1,2,3]}}')
```
This is not a JSON read bug, instead we need a way to compare strings
extracted from JSON inputs in a manner that isn't susceptible to
immaterial whitespace differences.

For more please refer to
https://github.com/NVIDIA/spark-rapids/issues/11154 .

Signed-off-by: Firestarman <firestarmanllc@gmail.com>