[BUG] Auto compact tests should use the confs with auto compact enabled (#13010)
`test_auto_compact_basic` and `test_auto_compact_partitioned` in
`test_auto_compact_min_num_files` are using the `_conf` object for the
writing which does not have the auto compact enabled. As such, none of
them actually performs the auto compact. These tests are passing because
they do not validate whether the table written is optimized or not. This
PR fixes this problem.

Signed-off-by: Jihoon Son <ghoonson@gmail.com>