Support RuntimeReplaceable version of StructsToJson for Spark 400 [databricks] (#13029)
contributes to https://github.com/NVIDIA/spark-rapids/issues/12941

## Analysis
From Spark 400, it transforms some expressions to `Invoke` with
`Literal` and `XxEvaluator`, e.g.:
StructsToJson => Invoke(Literal(StructsToJsonEvaluator), "evaluate",
type, arguments)

Please search with `override def replacement: Expression = Invoke`, and
gets:
```scala

  override def replacement: Expression = Invoke(
    Literal.create(evaluator, ObjectType(classOf[StructsToJsonEvaluator])),
    "evaluate",
    dataType,
    Seq(child),
    Seq(child.dataType)
  )

  override def replacement: Expression = Invoke(
    Literal.create(evaluator, ObjectType(classOf[ParseUrlEvaluator])),
    "evaluate",
    dataType,
    children,
    children.map(_.dataType))

  override def replacement: Expression = Invoke(
    Literal.create(evaluator, ObjectType(classOf[SchemaOfCsvEvaluator])),
    "evaluate",
    dataType,
    Seq(child),
    Seq(child.dataType),
    returnNullable = false)

  override def replacement: Expression = Invoke(
    Literal.create(evaluator, ObjectType(classOf[SchemaOfJsonEvaluator])),
    "evaluate",
    dataType,
    Seq(child),
    Seq(child.dataType),
    returnNullable = false)

  override def replacement: Expression = Invoke(
    Literal.create(evaluator, ObjectType(classOf[XPathEvaluator])),
    "evaluate",
    dataType,
    Seq(xml),
    Seq(xml.dataType))
```

The `XXEvaluator` e.g. `StructsToJsonEvaluator` is not a sub-class of
Expression.
The `Literal` can not be convert to `GpuLiteral`, becasue the `Literal`
in `Invoke` is just a Java class referance to `XXEvaluator`, it's not a
real literal and can not generate a column.
It's hard to add a `GpuInvoke`.

### Solution
Add a rule to do the revert of the above transforms before
`GpuOverrides`.`applyOverrides`
After the revert, it's easy to reuse the existing code, and do not need
to add a `GpuInvoke`.

### Solution 2 [not use]
Reuse the checks: e.g.: the check in StructsToJson ExprRule, how to
dispatch the check in Invoke ExprRule to sub types?
Handle the `disabledByDefault` for sub kinds of GpuInvoke.
Add GpuInvokeMeta: how to reuse the `tagExprForGpu` in the xxMeta e.g.:
GpuStructsToJsonMeta. We may need to create a temp GpuStructsToJsonMeta
and do tag, then copy the fallback infos to GpuInvokeMeta.
Add GpuXXEvaluator like Spark, this need to do refactor the existing
code.
`XXEvaluator` is not expression, `convertToGpu` is not suffient, so
should add code to conver `XXEvaluator` to GpuXXEvaluator.
Add GpuInvoke, call to GpuXXEvaluator
This solution is complicated.

### changes
-  Add RevertInvokeRule
- Fix test case `test_structs_to_json`, it can now run on GPU on Spark
400.
- Fix test case `test_structs_to_json`, explicity sepcify
`timestampFormat`:
GPU supports:     yyyy-MM-dd'T'HH:mm:ss[.SSS][XXX]
Spark 400 by default gets:   yyyy-MM-dd'T'HH:mm:ss[.SSS][XXXXX]

### Future PRs
Add more case paths in `RevertInvokeRule` to revert all kind of `Invoke`
expressions.

Signed-off-by: Chong Gao <res_life@163.com>

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>