Spark-4.0: Enable ansi mode for some of the integration tests [databricks] (#13043)
This PR contributes to
https://github.com/NVIDIA/spark-rapids/issues/11106

After revisiting all the tests with disable_ansi_mode label, the tests
in this PR can be enabled to run with ansi enabled for Spark-4.0. Most
of the tests were disabled due to lack of ansi support for SUM
aggregation. It was added recently and the tests pass now.

For `test_many_column_project` in parquet_test.py:
- The ProjectExec would fallback to CPU if the dataframe is read from
Scan ExistingRDD, so added allow_non_gpu for that test.


Testing:

Ran the integration tests on each file in this PR and there are no
failures:
 `./integration_tests/run_pyspark_from_build.sh -k '<PYTHON_FILE_NAME>`

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>