Revert "Support RuntimeReplaceable version of StructsToJson for Spark 400 [databricks] (#13029)" (#13075)
closes https://github.com/NVIDIA/spark-rapids/issues/13070
This reverts commit 4ff87790

https://github.com/NVIDIA/spark-rapids/issues/12941 PR introduced an
issue, it can not handle fallback cases.
PR #12941 rewrites `invoke(liternal(StructsToJsonEvaluator))` to
`StructsToJson`, if it does not fallback, it works; If it
fallbacks(e.g.: meets unsupported timestamp formats), it will run on CPU
and try to generate code and will fail.

Need to find another way to adapt `RuntimeReplaceable` expr.

Signed-off-by: Chong Gao <res_life@163.com>

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>