Update tests/aggregations for ANSI mode, when they don't care (#13032)
This contributes to https://github.com/NVIDIA/spark-rapids/issues/5114
but does not finish it. Most aggregations that we support (really
everything except SUM and AVG) don't care about ANSI support, so this
just updates the tests to enable them and for the bit aggs enables them
in the code too.

Some of the time I have it run with ANSI on and off for the tests, but
most of the time I just run the test with the default.

There are a few cases where a SUM would overflow or a test will hit
another ANSI error. In those cases I typically just force the test to be
non-ANSI.

Some of the tests depended on other expressions that are not updated for
ANSI yet (like multiply). So in those cases I kept the annotation to
skip the test is ansi is on by default and pointed to the issue for it

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>