Handle StagedTableCatalog when creating AppendDataExecV1 (#13083)
This introduces changes to handle StagedTableCatalog when creating
AppendDataExecV1 in the Spark RAPIDS plugin.

The main goal is to address #13018 when working with Spark's
StagedTableCatalog interface during append operations, ensuring that the
RAPIDS plugin correctly handles tables that are not yet fully committed
but are in a "staged" state.

The code now explicitly checks for and handles instances where the
catalog is a StagedTableCatalog when creating the AppendDataExecV1
execution plan

fixes #13018

---------

Signed-off-by: Raza Jafri <raza.jafri@gmail.com>