Revert "declare shuffle thread for threaded writer and reader" (#13056)
This reverts https://github.com/NVIDIA/spark-rapids/pull/12973 for the
25.08 release.

We have seen this change cause regressions that relate to locking within
the SparkResourceAdaptor in spark-rapids-jni. We would like to hold off
on the change for now. The main effect of reverting is we go back to
having a higher chance of OOM false positives => added spill pressure.

There's a second report of CI failing after this change, and this change
may clear that too, but we want to get to the bottom of the CI issue
separately.

I have filed https://github.com/NVIDIA/spark-rapids/pull/13056 for
follow up after this PR

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>