Fix up the exception type for casting strings to decimals. (#13089)
close https://github.com/NVIDIA/spark-rapids/issues/10908

This PR catches the cast exception from the JNI and throws an exception
the same as Spark, and adds in some tests for it.

This PR depends on https://github.com/NVIDIA/spark-rapids-jni/pull/3521
which really fixes the bug in the linked issue.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>