Use the original input column to get the overflow decimal values (#13109)
close https://github.com/NVIDIA/spark-rapids/issues/13065

This PR changes to use the original input column to get the overflow
decimal values to build the exception message instead of the rounded
column when casting decimals to decimals. Because the rounded column may
have decimals with higher precision (e.g. 4) than the one (e.g. 3) of
the original column type, leading to the error in the linked issue.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>