Add in code to halt on shutdown timeouts (#13031)
We have run into cases where an executor gets stuck on shutdown. We do
not know the root cause of it yet, but this should mitigate the issue
where we can try and ask the process to shut down nicely, but if it
still has things up and running after a set timeout (currently 40
seconds), then we will call Runtime.halt, which is similar to a kill -9
for the process.

I think we may want to add some debugging before we call halt, but I
want to be careful with it because I don't want it to get in the way of
us shutting down so suggestions are welcome.

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>