Fallback to CPU for try_ functions [databricks] (#13095)
This PR contributes to
https://github.com/NVIDIA/spark-rapids/issues/11884

From Spark-3.4, some of the try_* functions such as try_add, try_divide
and others don't call TryEval for evaluating the expression result.
Instead, a new argument EvalMode.TRY is passed to the operator case
class.
Currently, we don't differentiate between EvalMode.ANSI and
EvalMode.TRY. So we end up throwing error for try_* functions.
In this PR, we fallback to CPU for the functions where we don't handle
it correctly.
Checked the Apache Spark-4.0 source code and tested all the functions
which have try_* prefix.
Added only few tests to verify that the operators indeed fallback to CPU
for try_* functions.

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>
Co-authored-by: Chong Gao <gaochong.gc@qq.com>
Co-authored-by: Liangcai Li <firestarmanllc@gmail.com>