Support side-effect check for the "else" path in GpuCaseWhen (#13149)
close https://github.com/NVIDIA/spark-rapids/issues/13137

This PR adds in side-effect check to the "else" path in `GpuCaseWhen`.
And the `hasSideEffects` should be always true for `GpuElementAt`
because it will throw out exceptions whenever there is 0 index, even the
ansi mode is off.

Before this PR, `GpuCaseWhen` will not run into the evaluation with
side-effect if only the "else" path has side-effect. Then the "else"
expression will always be evaluated, leading to the error in the linked
issue.

The test case is simplfied from the example in the linked issue.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>
Co-authored-by: Chong Gao <gaochong.gc@qq.com>