Fix failed to_json test cases (#13162)
closes #13152

Restores proper fallback behavior for non-UTC timezone scenarios that
was inadvertently removed in PR
https://github.com/NVIDIA/spark-rapids/pull/13126

### Bug analisis
The test cases update in
[PR](https://github.com/NVIDIA/spark-rapids/pull/13126/files#diff-e43717af9471ebedcd992bac98ab4b21fb7a771205fbcbc0760446f283a6d5f0)
introduced this bug.

```diff
-structs_read_allowed_non_gpu = non_utc_project_allow if is_before_spark_400() else \
-    ['ProjectExec', 'Invoke', 'Literal']
```
The PR #13126 supports: ['ProjectExec', 'Invoke', 'Literal'] for Spark
400+, but in the above diff, the `non_utc_project_allow` was also
removed. For all the Spark versions, it will fallback if timezone is
non-UTC.

### Fix
Mark fallback is allowed when timezone is non-UTC.

Signed-off-by: Chong Gao <res_life@163.com>

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>