Host limits fix device count (#13158)
fixes https://github.com/NVIDIA/spark-rapids/issues/13040

(this branch is also based on another PR branch,
https://github.com/NVIDIA/spark-rapids/pull/13157)

Spark RAPIDS jobs are often run on hosts with only one GPU. However,
it's possible to run on a host with multiple GPUs, but the plugin will
only use one GPU within an executor. If we see multiple GPUs on the host
during initialization, we want to be conservative and assume some other
process, whether it's another executor in the Spark job or just some
entirely different workload, is running and using the other GPUs, and
presumably would also need a proportional amount of the available host
memory. Therefore, we divide the total amount of host memory we see by
the number of GPUs.

This PR addresses an issue where the unit tests were all assuming the
host has only one GPU. These tests would then pass if that's true but
fail if there are multiple devices (a workaround is to set
`CUDA_VISIBLE_DEVICES` to one device id, but regardless it won't be
necessary after this fix). Instead of checking the actual device count
on the host as in a real query execution, the tests would now just
inject values in order to be more hermetic and also test different
scenarios.

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>