Allow for better parquet type conversion (#13165)
This fixes #11512

It adds in support for wider type conversion in parquet and matches what
Spark 4.0.0 supports. We didn't make it backwards error compatible so we
might be able to support some data type conversions on older versions of
Spark that the CPU does not support.

On a side note to make this work I needed to fix some issues with
data_gen and how TimestampNTZ works as it does not work for all versions
of pyspark and does if you tried to use it you would get errors even if
you tried to skip the test.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>