Add tests for the cases of decimal precision going beyond 38[databricks] (#13180)
close https://github.com/NVIDIA/spark-rapids/issues/10515

The Spark PR (https://github.com/apache/spark/commit/7120e6b88f2) has
mainly updated the computation of result decimal type for decimal
relevant operations during the analysis phase when no GPU things are
involved. It affects only the case of the precision of the result
DecimalType goes beyond the MAX_PRECISION(38). And a `Cast` will be
inserted (for example:
[decimalAndDecimal](https://github.com/apache/spark/blob/branch-4.0/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/DecimalPrecisionTypeCoercion.scala#L141),
[widenTypes](https://github.com/apache/spark/blob/branch-4.0/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/TypeCoercionBase.scala#L223))if
the change results in a different decimal type from the original one for
a column.

So there seems to be no action for us, but this PR still adds in some
tests to confirm this.

Verified locally on Spark 400, the tests can always pass.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>