Run try_ functions aggregation arithmetic tests on single partition (#13196)
This fixes https://github.com/NVIDIA/spark-rapids/issues/13179

The test_try_sum_groupby_fallback_to_cpu was intermittently failing in
nightly builds, showing different results between "CPU" and "GPU" runs
despite both correctly falling back to CPU execution. So the CPU results
were not deterministic.
The overflow detection on aggregations is inconsistent for the CPU.
These functions use sequential addition at each row with overflow
checking so the result depends on partitioning and the order of the data
being processed.
In this PR, we force the data to be on a single partition so that the
results are always deterministic as the updateExpression step happens in
a single partition and the mergeExpression has the same data for all the
runs.

Updated the other try_ arithmetic tests with num_slices since the above
issue can happen on the both groupBy queries and reduction queries.

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>