Fix expression combining when only some can combine (#13213)
This fixes https://github.com/NVIDIA/spark-rapids/issues/13204

The code is kind of complicated aka a bit of a hack, with how it finds
expressions that can be combined together. It works and it is this way
for performance. I simply made the code that gets the replacement return
it optionally so that we can then not replace it if we should not.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>