Fix case where null start index or length would throw invalid parameter, instead of returning null rows (#13236)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13164

This fixes implementation and test issues with `array_slice` where we
were not taking into account null columns or scalars in the `list`,
`start` arguments to `array_slice`. The test that failed CI
intermittently would do so when it expected the `array_slice` function
to fail when `length` was negative, but it didn't, instead returning
null rows. The reason for this was because either the rows in `list` or
`start` could be null, and in these scenario spark returns null for the
result of slice, and doesn't throw.

The fix is to check for these rows, and we use a function to "merge
nulls" onto the `length` column, if length is not a scalar. We do this
because there could be nulls at precisely the rows where length is
negative, and therefore no exception should be thrown.

The patch adds tests to go through more edge cases than before.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>