Close shared buffers at batch read completion in kudo deser stream iter (#13248)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13145

Currently we allocate shared buffers in the deser iterators as an
optimization. The issue is we wait until the task completes to close
these buffers, but a single task can be associated with a large numbers
of these iterators. This change closes the buffers as soon as we see
that we are finished with a batch.

Signed-off-by: Zach Puller <zpuller@nvidia.com>