Add override tag to the 'ensReqDPMetricTag' and change it to a 'val'[databricks] (#13253)
This is a follow-up fix for
https://github.com/NVIDIA/spark-rapids/pull/13243.

It also fixes https://github.com/NVIDIA/spark-rapids/issues/13254 to
pass the premerge, along with some changes for the error message
updates.

'ensReqDPMetricTag' turns out to be a 'val' of Scala and requires to be
overrided.

Verified it inside the driver node of a Azure cluster, build has passed.
IT is running ...

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>