Print DB runtime version details for debugging purpose [skip ci] (#13261)
### Description
This PR is to update databricks build and test script to print DB build
info for debugging purpose.

### Checklists
This PR has:

- [ ] added documentation for new or modified features or behaviors.
- [x] updated the license in the source code files when it is required.
- [ ] added new tests or modified existing tests to cover new code
paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)

Please select one of the following options:
- [ ] Performance testing has been performed and its results are added
in the PR description.
- [ ] An issue is filed for performance testing and its link is added in
the PR description. (Select this if performance testing will not be
completed before the PR is submitted.)

Signed-off-by: Yinqing Hao <haoyinqing@gmail.com>