Fix fallback test params for Delta MergeCommand, UpdateCommand, and DeleteCommand (#13280)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13279.

### Description

We have integration tests that test the fallback for merge, update, and
delete commands with various configs. One of the cases tested is the
"default" config when there is no config set for these commands. As
these commands are enabled by default now, the fallback tests are
failing for this case. Since the default behavior is changed only for
spark 3.5.3+, the test parameters are now set differently based on the
spark version so that the default case won't be tested for spark 3.5.3+.

### Checklists

This PR has:

- [ ] added documentation for new or modified features or behaviors.
- [ ] updated the license in the source code files when it is required.
- [x] added new tests or modified existing tests to cover new code
paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)

Please select one of the following options:
- [ ] Performance testing has been performed and its results are added
in the PR description.
- [ ] An issue is filed for performance testing and its link is added in
the PR description. (Select this if performance testing will not be
completed before the PR is submitted.)

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>