Fix the checklists in the PR template to reduce confusion [skip test] (#13320)
### Description

This PR fixes two issues.

- The checkbox for the updating license is redundant as we now have a CI
job running to test it.
- The checkboxes for performance testing have been merged into one and
placed together with other items. Previously, the separate instruction
and section for performance testing caused some confusion as if it was
enforcing to choose between two options, either performance testing has
been done or performance testing will be done later. It was not clear
that there is a third option for when the performance testing is not
necessary.

### Checklists

This PR has:

- [ ] added documentation for new or modified features or behaviors.
- [ ] updated the license in the source code files when it is required.
- [ ] added new tests or modified existing tests to cover new code
paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)

Please select one of the following options:
- [ ] Performance testing has been performed and its results are added
in the PR description.
- [ ] An issue is filed for performance testing and its link is added in
the PR description. (Select this if performance testing will not be
completed before the PR is submitted.)

Signed-off-by: Jihoon Son <ghoonson@gmail.com>