Add support for optimize table command for Detla IO 3.3 [databricks] (#13313)
Fixes #13012.

Note: the first commit of this PR is completely written using Copilot +
GPT-5 with VS code. It is impressive that it generated code that
compiles and passes tests successfully. Subsequent commits are for
cleaning up the code, removing a redundant test, and adding a missing
test.

### Description

We currently support auto compaction and optimize write, but we don't
support the manual compaction using the optimize table command. This PR
adds support for the optimize table command for manual compaction.

The new optimize table command is just an interface for users and share
the core logic of compaction with auto compaction, which is implemented
in `GpuOptimizeExecutor`. As the compaction is well-tested in
https://github.com/NVIDIA/spark-rapids/issues/12996 including
performance testing, the optimize table command is enabled by default.

Implementation-wise, there is a slight difference in the class location
between the plugin and Delta IO. In Delta IO, `OptimizeExecutor` and
`OptimizeTableCommand` locate in the same file
`OptimizeTableCommand.scala`. But in the plugin, they are in
`GpuOptimizeExecutor.scala` and `GpuOptimizeTableCommand.scala`,
respectively. This is because `GpuOptimizeExecutor` was added in
https://github.com/NVIDIA/spark-rapids/pull/13023 first, and
`GpuOptimizeTableCommand` is added in this PR. I'm not sure if we should
merge these files so that it would match to the structure in Delta IO.
I'd happy to do it if someone thinks we should.

The new command is tested by the new tests added in
`delta_lake_optimize_table_test.py`.

### Checklists

This PR has:

- [ ] added documentation for new or modified features or behaviors.
- [ ] updated the license in the source code files when it is required.
- [x] added new tests or modified existing tests to cover new code
paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)

Please select one of the following options:
- [x] Performance testing has been performed and its results are added
in the PR description.
- [ ] An issue is filed for performance testing and its link is added in
the PR description. (Select this if performance testing will not be
completed before the PR is submitted.)

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>