add nvtx, fix row count, add data size metric (#13250)
### Description


This PR adds nvtx ranges to the KudoGpuSerializerInstances for
performance analysis, and fixes bugs in the kudo gpu code where we were
failing to write certain bits of metadata which are needed by spark to
do AQE correctly. Namely, the following bugs were fixed:
1. The ColumnarBatch row counts were being set incorrectly. They were
being set to the cumulative partition indexes instead of individual
partition sizes, eg. if the sizes are 1, we should set 1, 1, 1, 1, but
were setting 1, 2, 3, 4.
2. We were failing to write the shuffle output datasize spark metrics,
which was needed by AQE.

### Checklists


This PR has:

- [x] added documentation for new or modified features or behaviors.
- [x] updated the license in the source code files when it is required.
- [x] added new tests or modified existing tests to cover new code
paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)

Please select one of the following options:
- [ ] Performance testing has been performed and its results are added
in the PR description.
- [x] An issue is filed for performance testing and its link is added in
the PR description. (Select this if performance testing will not be
completed before the PR is submitted.)

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>
Co-authored-by: Alessandro Bellina <abellina@gmail.com>