Fix test_delta_optimize_fallback_on_clustered_table test failure with Spark 3.5.6 [databricks] (#13333)
Fixes #13331.

### Description

It is a [known issue](https://github.com/delta-io/delta/issues/4671) in
delta that `CREATE OR REPLACE TABLE` sometimes fails with the error
`Table {name} does not support truncate in batch mode`. The test in
question in fact does not need to use `CREATE OR REPLACE TABLE`. It is
left over after some refactoring. I changed it to `CREATE TABLE`
instead. I manually verified this change in my workstation with Spark
3.5.6 and Delta 3.3.0.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Jihoon Son <ghoonson@gmail.com>