Add Rapids file io layer. (#13252)
Closes  #13138 

### Description

This pr adds a `RapisFileIO` layer to provide abstraction for differnt
underlying storages. Also in this pr we provide two implementatiosn:

1. `HadoopFileIO` to be compatible with existing behavior.
2. `IcebergFileIO` so that we could use `iceberg.FileIO` to access
iceberg parquet data, which may contains security information and
encryption keys.

Changes in this pr is refactoring of file system, so existing test has
covered changes in this pr.

### Checklists


This PR has:

- [x] added documentation for new or modified features or behaviors.
- [x] updated the license in the source code files when it is required.
- [x] added new tests or modified existing tests to cover new code
paths.
     
Please select one of the following options:
- [ ] Performance testing has been performed and its results are added
in the PR description.
- [ ] An issue is filed for performance testing and its link is added in
the PR description. (Select this if performance testing will not be
completed before the PR is submitted.)
- [x] This issue is mostly refactoring and being compatible with
existing behavior, so we could skip perf test.

---------

Signed-off-by: liurenjie1024 <liurenjie2008@gmail.com>