This fixes a big regression due to serialization of the hadoop configuration (#13354)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13353

### Description

This fixes a 46% performance regression on NDS @ sf3k in our performance
clusters due to this change:
https://github.com/NVIDIA/spark-rapids/pull/13252. The culprit was a
hadoop configuration that was serialized with the body of the task.

The regression is fixed by accessing the broadcast directly at runtime
from the task.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>