Add read tests for the delta clustered table [databricks] (#13327)
Fixes https://github.com/NVIDIA/spark-rapids/issues/12552.

### Description

We currently have no test for reading from clustered tables for delta.
This PR adds some test to cover it.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>