Add rest catalog integration tests for iceberg (#13329)
Fixes #13301. 

### Description

This pr adds integration test for iceberg interacting with rest catalog.
Its setup is as following:

1. It setups a rest catalog implemenation using
[lakekepeer](https://docs.lakekeeper.io/)
2. It setups s3 compatible storage using minio

All theses are setup using docker compose.

It's a test only change so we don't need to exercise perform test.

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
     
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Ray Liu <liurenjie2008@gmail.com>